requirements.txt
langgraph>=0.0.1
langchain>=0.1.0
langchain-core>=0.1.0
google-generativeai>=0.3.0
google-api-core>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
